+++ CHANGES.md
@@ -15,19 +15,20 @@
 [Unreleased](https://github.com/bird-house/birdhouse-deploy/tree/master) (latest)
 ------------------------------------------------------------------------------------------------------------------
 
-[//]: # (list changes here, using '-' for each new entry, remove this when items are added)
+## Changes
+- Allow users to submit a Weaver job requesting to store outputs to the public location instead of their user-workspace.
 
 [1.33.5](https://github.com/bird-house/birdhouse-deploy/tree/1.33.5) (2023-10-02)
 ------------------------------------------------------------------------------------------------------------------
 
-## CHANGES
+## Changes
 
 - Adding a description for the STAC service that will be served at the `/services` endpoint
 
 [1.33.4](https://github.com/bird-house/birdhouse-deploy/tree/1.33.4) (2023-10-02)
 ------------------------------------------------------------------------------------------------------------------
 
-# Fixes
+## Fixes
 - Clean up: Make bind-mount locations more flexible
 
   Clean up unused variables and correct file paths from the changes made in 1.33.2